refactor(bridge): address Plan 1 code-review — braces, comments, guards
Standards / spec review of the Plan 1 branch (both axes). Applied:

- Braces on all new control flow in doGet's `selector=backend` block and
  doPan's span/rate — AGENTS.md C++ Style Guide ("Braces on all control
  flow"). The copied `get hostnb` neighbour omits them; the new code does not.
- doGet's `selector=backend` non-property return now echoes the `selector`
  variable instead of a hardcoded "backend" literal (matches the property path).
- dsp.get in Hl2Backend::invokeExtension: a caller-affinity guard around the
  BlockingQueuedConnection — call the read lambda directly when already on
  m_ioThread, so a future I/O-thread caller cannot self-deadlock (the
  publishIoDsps() guard shape).
- AudioMetrics::combSpacingHz doc + tests.cmake comment: "cepstral" ->
  "spectral autocorrelation" — the implementation stopped being a cepstrum
  when the weak first cut was replaced.
- capture_audio analyze `frames` is `int`, not `double` — matches the plan's
  stated shape and the sibling probe results.
- automation_pan_perf_verbs_test: the `get dsp selector=backend` row now
  asserts an informative ok:false decline (any non-empty error) rather than a
  specific string — a non-HL2 backend's decline text is its own.

Deferred (noted in the plan): extracting the 3rd copy of the synchronous-
extension scaffold shared with `get hostnb` (touches shipping code); a
standalone automation_dsp_readback_test with a connected Hl2Backend fixture
(same gap nb.get has — hardware-covered in Plan 5).

## Test

Rebuilt aethercore + AetherSDR clean. ctest in the CI container:
audio_metrics_test, hl2_dsp_readback_test, automation_audio_analyze_test,
automation_pan_perf_verbs_test, bridge_docs_check, bridge_docs_subaction_audit,
wdsp_channel_test, hl2_rxdsp_test, hl2_noise_blanker_test, hl2_notch_seed_test,
hl2_dbref_test — 11/11 passed.

[built in CI container; unsigned — no signing key in this environment]

+++ src/core/AudioEngine.cpp
@@ -3824,7 +3824,7 @@ QJsonObject AudioEngine::automationAudioCaptureAnalyze(const QString& tap) const
     return QJsonObject{
         {QStringLiteral("ok"), true},
         {QStringLiteral("tap"), want.isEmpty() ? QStringLiteral("all") : want},
-        {QStringLiteral("frames"), static_cast<double>(mono.size())},
+        {QStringLiteral("frames"), static_cast<int>(mono.size())},
         {QStringLiteral("sampleRate"), sampleRate},
         {QStringLiteral("channels"), channels},
         {QStringLiteral("dominantHz"), tones.size() > 0 ? tones[0].hz : 0.0},
